:sparkels: add ImageOptimize component
